(IAC-646) Update builder to leverage pwshlib
Prior to this commit the builder specified pwshlib in the metadata
for dependencies but did not specify it as a fixture.

This commit ensures pwshlib will be added as a fixture for testing
and development. It also updates the writing of the metadata and
fixture files to write UTF8 without a BOM to prevent parsing errors
during Puppet runs.