(IAC-651) Improve Module Handling
Prior to this commit the build script modified the Module Path
before attempting to import/use EPS. Further, the check for EPS
did not look for it on the system, only in the session.

This commit swaps the ordering so that EPS is sought, installed
if needed, and loaded prior to modifying the PSModulePath. It
also updates the search.