(IAC-651) Pull from Gallery instead of CSV
Prior to this commit the builder consulted a local CSV file
to determine which PowerShell modules with DSC resources to
puppetize. This commit replaces the implicitly passed CSV
with three parameters:

- PowerShellModuleName: The name of the module to pull from
  the gallery. Defaults to PowerShellGet for now.
- PowerShellModuleVersion: The version of the module to pull
  from the gallery. If unspecified, will pull latest.
- PuppetModuleName: The name for the output Puppet module.
  If unspecified, will default to the PowerShell module
  name, downcased.

In the process of doing so, this commit also updates the
vendoring logic, moving the Puppet module build phase (via
the pdk) to an earlier step and ensuring the vendored
PowerShell modules are placed in the appropriate folder
so they can be discovered during a Puppet run - for now,
the vendored resources live in lib/puppet_x/dsc_resources.