(IAC-651) Improve Type Generation
This commit heavily updates the type generation template and
the resulting generated types. It leverages a new PowerShell
function, Get-DscResourceTypeInformation, which introspects
a DSCResourceInfo object (as supplied by Get-DscResource) to
return:

- The Name of the DSC resource, downcased
- The friendly name
- The provider name
- The resource type information
- the resource module, name, and version
- How the resource is implemented (PowerShell, binary)
- The relative MOF path, if any
- Which properties may be called via Invoke-DscResource
- Which properties may be called via the DSC DSL
- ParameterInfo, which includes the following info:
  - The name of the parameter, downcased
  - The default value for the parameter
  - The type of the parameter
  - Whether or not the parameter is mandatory for get
    method invocations
  - Whether or not the parameter is mandatory for set
    method invocations
  - the mof_type for the parameter
  - whether or not the parameter has an embedded instance.

It retrieves this information by combining the information
available on the DSCResourceInfo object with information
obtained by parsing the AST of its source, if the resource
was implemented in PowerShell (this function does not parse
binary or python implementations, for example).

For determining whether or not a parameter is mandatory,
this function introspects the Get-DscResource and the
Test-DscResource functions in the AST; if these cannot
be found (because the resource in binary, not written in
PowerShell, or class-based) it instead relies on the
Required property of that parameter as surfaced in the
allowed_properties of the DSCResourceInfo object.

If the paramter *is* marked as mandatory for the get
method (either in the function or via the Required
property), then when the type information for that
parameter is translated as an attribute it will also
be given the namevar behavior. This will enable us to
implement the simple_get_filter feature of the resource
api in a future commit.