(MAINT) Refactor create/update/delete
Prior to this commit the implementation for the create, update,
and delete methods was half-done and would be repeated across
all three. As they share the same functional implementation -
turn the should hash into a resource passable to the underlying
Invoke-DscResource call with the Set method - this commit pulls
the logic for calling Invoke-DscResource with the Set method,
including its return information, and places it into a ruby
helper function, invoke_set_method, which the other methods can
then call.