(IAC-651) Ensure the generated resources are idempotent
Prior to this commit the get method in the provider was not
fully implemented and returned mock data. This meant that the
resource_api would never be able to find the resource on a
target node to compare to and would always call the create
method, even if it already exists and is in sync.

The DSC API does not have an equivalent to the Puppet
provider method instances and therefore can not be used to
return a list of all the resources of that type on the
target node.

Instead, implementing the simple_get_filter feature enables
us to look for *each* resource by its mandatory parameters
(marked as namevars) using the Get method of a call to the
Invoke-DscResource command.

This means the Puppet will be able to tell from the resulting
data hash whether or not a resource exists on the target node
(and call create or delete as needed) and, if it does exist,
whether or not the properties being managed by Puppet are in
sync - and, if they are not, it will report which properties
were not in sync, their values, and the values they are being
updated to; this is functionality not available in normal DSC
implementations.

This *also* finally makes the provider implementation
idempotent, as it no longer always calls create regardless of
the state of the resource and does exactly and only what is
required given the actual state on the target node.