(IAC-650) Move Get-DscResourceTypeInformation
This commit moves the function Get-DscResourceTypeInformation
into the public functions folder of the newly scaffolded
module and updates the dotsource in the build script with an
import of the module.

From this point forward as functionality is decomposed from
the build script, so long as it is appropriately exported
from the module it will be available.

This commit also makes minor updates to the File Integrity
tests as they do not yet support UTF8 without a BOM. As we
need to interop with Ruby, any and all BOMs should be avoided.