(MAINT) Improve Get-DscResourceTypeInformation params
This commit overhauls the parameter declarations for the public function,
Get-DscResourceTypeInformation, ensuring that it has two parameter sets
to distinguish between being passed a full DSC Resource Info object and
having to query the API for one based on Name (and, optionally, module).

It also ensures that the types and names for the parameters match those
specified by the Get-DscResource command precisely. It now allows a
user to pass an array of objects with name and module properties to
retrieve information.

Finally, this commit updates the Process block to ensure that it can
handle being passed multiple DSCResource names *not* from the pipeline.