(IAC-667) Add Get-DscResourceTypeInformation unit tests
This commit refactors the Get-DscResourceTypeInformation function
slightlyt to be more testable, as Add-Property is mockable and
an assignment is not.

It also adds a first pass at unit tests for the function and
its dependent private function, Get-DscResourceParameterInfo.

It slightly reworks the DSC Resource information retrieval in the
process block to ensure that it can properly handle being passed
multiple objects which are *not* of the DscResourceInfo type
without breaking, assigning the resources for processing  to a
teemp variable instead of overwriting the parameter variable.