(MAINT) Ensure custom types are autoloaded on module import
Prior to this commit the custom types implemented for returning formatted data
with Get-DscResource were loaded manually in the Build script, not automatically
on module import.

This commit moves the type definition information into the module, ensures that
it will be automatically loaded when the module is imported, and removes the logic
for manually updating the types from the build script.