(IAC-667) Add tests for Get-DscResourceParameterInfo
This commit adds two basic integration tests for the private function
Get-DscResourceParameterInfo:

1. What the function will return if it can parse the AST;
   this test loads the Archive DSC resource at v2.12.0.0,
   ensuring that the test output should always be static.
2. What the function will return if it cannot parse the
   AST; this test loads the File DSC resource which is
   included in-the-box without a version as a binary DSC
   resource.

This is done because the DscResourceInfo object was, in
the end, impossible to mock for testing. That left us
with needing to pass real DscResourceInfo objects to the
function or to stop strongly typing the function input;
we chose to pass real objects to reduce testing and design
complexity at this stage.

As we may return to this function in the future and parse
via CIM instances, it does not seem worth it to invest
heavily in a rewrite *now*.