(IAC-668) Replace templating engine
Prior to this commit the builder relied on the EPS templating
engine which has not been actively maintained for several years
now. This made the builder fragile and added cognitive load.

This commit updates the builder to do away with the templating
engine completely, opting instead to implement a file copy for
the static files and the use of here-strings in public functions
to retrieve the text for a Puppet Resource API compliant type
and provider for each DSC resource.