(MAINT) Fix parameter type munging
Prior to this commit the munging for the type of a
DSC parameter improperly dropped the final ] of a
type like Optional[String["foo", "bar] - this made
the validation of the type fail.

Further, it inserted keys into this nested structure
wrapped in single quotes instead of double quotes,
which also broke the type validation.

This commit corrects the edge case.