(IAC-666) Implement ConvertTo-PuppetResourceApi
Prior to this commit the conversion of a DSC resource into a Puppet
resource required numerous steps, all of which were exposed as public
functions but which were not, on their own, useful.

This commit exposes as a public function, ConvertTo-PuppetResourceApi,
which utilizes underlying functions in the module to retrieve the
parameter information of a DSC resource and output a well-formed string
for both the type and provider files to be written to disk.