(IAC-666) Privatize functions
Prior to this commit the Get-DscResourceTypeInformation,
Get-ProviderContent, and Get-TypeContent functions were
erroneously exposed as public functions in order for the
build script to function.

This commit correctly privatizes those functions which
are not useful on their own now that the more useful
ConvertTo-PuppetResourceApi function exists, which they
were written to enable.