(MAINT) Ensure friendly name is written to type
Sometimes the friendly name for a DSC resource is not defined;
in these cases the friendly name be _something_ other than null.
We default to the name of the DSC resource; after reviewing all
of the DSC resources available on the Gallery, every resource
that has a friendly name has one that matches the resource name.