(IAC-710) Remove types.ps1xml dependency
Prior to this commit the module relied on the Types.ps1xml file to
munge and alias DSC Resource property information to make processing
easier in other functions. This was useful but at the cost of being
very hard to reason about, review, update, and also impossible to
usefully test with the given tools.

This commit removes our reliance on the types file entirely, moving
the relevant functionality into private functions and updating the
logic elsewhere to be more transparent.

It creates the following private functions:

- Test-EmbeddedInstance, which does a best-guess at whether or not
  a particular PowerShell data type represents an embedded cim
  instance or not. As of the time of this commit, no PowerShell
  modules were found with data types not covered by this function
  which were not actual embedded instances.
- Get-ShortType which strips a DSC Resource Property PropertyType
  of the outer brackets, making it easier to re-insert later.
- Get-PuppetDataType which does a best-effort match of PowerShell
  data types to Puppet data types. Unfortunately, we have not been
  successful in finding a programmatic way to parse all of the
  embedded instances discovered on the gallery. We have a potential
  option which may help us discover embedded instance structures for
  class-based resources, but not mof-based.

  Note that although we cannot build the struct for *every* embedded
  instance type we **have** manually added the struct for PSCredential
  objects as these are extremely common.

  This also fixes some of the data types which were not properly parsed
  by the Types.ps1xml file, ensuring that the data types are actually
  mapped instead of just writing PowerShell data types in the Puppet
  Resource API type file.

There remains some room for improvement, particularly around discovering
reference documentation and handling embedded instances.