(IAC-707) Add Update-PuppetModuleReadme
This commit takes an initial pass at adding an automatically gnerated
readme file to the Puppetized module, covering basic usage and clarifying
that the module is autogenerated, as well as pointing readers in the
right direction for troubleshooting, filing issues, and investigating.

The troubleshooting section is left a little bare as the bulk of this
documentation is scheduled to be written in IAC-716.

This readme should be updated as additional documentation is written
to reflect best practices and functionality.