(IAC-708) Replace ERB with ruby methods and document all methods
This commit reworks the definition of the PowerShell script to be
executed for invoking DSC resources to replace defining the params
hash in an ERB file with a ruby method, invoke_params.

In the process of doing so, manual testing showed that the current
implementation had some trouble handling both PSCredentials and
embedded CIM instances. This commit therefore also includes methods
to address these problems and improve debuggability of both the
provider behavior itself as well as the PowerShell script.

The new prepare_credentials and prepare_cim_instances methods can
be called to retrieve all definitions of either credential objects
or CIM instances from a given resource and return PowerShell code
to instantiate them and assign them to a unique variable which
can be interpolated into the eventual hash later.

This means the definition of each credential and CIM object is
performed separately and can be verified from the PowerShell script,
but also cleans up the definition of the InvokeParams hash and
prevents us from having to manage nested creation of instances and
hashes alike.

As we're doing away with the ERB template, it made sense to convert
the preamble and postscript into PowerShell scripts - which will also
allow us to test them in the future.

Both the preamble and postscript were modified to perform munging
on the return output of invoking with the Get method, which returns
a CIM Instance object represeting the DSC resource. It is easier
to munge this return object in PowerShell, via the new function,
Canonicalize-Result, than to munge the hash in Ruby.

Finally, this commit adds reference documentation to each of the
ruby methods, thereby reesolving IAC-708.