(MAINT) Ensure puppetized module versions are valid
Prior to this commit the versioning for the built Puppet module included
the fourth digit as is often included in PowerShell module versions.

Puppet tooling does not support four-digit versions.

This commit adds a new private function, Get-PuppetModuleVersion, which
can be used to translate a valid PowerShell module version into a valid
Puppet module version.

It does so by using the prerelease convention, mapping the fourth-digit
of a PowerShell module version to the first digit of the prerelease tag.

It further adds the ability to distinguish between rebuilt versions of
an output module where the source is the same but the output is different
because the builder was updated to fix a bug in type generation.

This means that the Puppet module version for a PowerShell module will
look like these examples:

- 1.2.3   becomes 1.2.3-0-0
- 1.2.3.4 becomes 1.2.3-4-0
- 2.3.21  becomes 2.3.21-0-0, if rereleased it is then 2.3.21-0-1

This keeps the versions human-readable and mappable between the two
packaging standards and also makes version comparison logic on our
end easier.