(IAC-671) Add New-PuppetDscModule function
This commit exposes a new public function for the module,
New-PuppetDscModule, which takes all of the functionality
previously living in the build script and turns it into
a callable function.

This commit also adds testing for this process, including
a wrapping the PSModulePath munging into a private function
for testing purposes since direct variable assignments can
not be mocked.

This public function allows a module user to specify a
single command and Puppetize a PowerShell module with DSC
resources and is the main interface for the module.