(IAC-671) Privatize internal functions
This commit privatizes several functions which are not useful in
isolation, only during a module Puppetization:

- Add-DscResourceModule
- Initialize-PuppetModule
- Update-PuppetModuleFixture
- Update-PuppetModuleMetadata
- Update-PuppetModuleReadme

These were more properly private functions but had to be exposed as
public functions earlier in development to keep the build script in
working order as it lives outside of the module.

With the implementation of New-PuppetDscModule, these functions can
be safely privatized, reducing command clutter for end users.