(MAINT) Check stdout for pdk success messages
Prior to this commit the Invoke-PdkCommand private function
only checked the stderr logs for a PDK command to verify
whether or not it was successful.

Since the PDK does not consistently send information to
stderr or stdout indicating successful completion of a
command, this commit updates the success filter to work
regardless of output stream.