(IAC-740) Add private functions for CIM parsing
This commit adds the Get-DscResourceParameterInfoByCimClass function,
an alternate implementation of Get-DscResourceParameterInfo, which
could not parse embedded CIM instances. The new function can, but
will only work if run as administrator.

A future commit will have to plumb this in, but for now the tests
are written and in place.

A current limitation of Pester is that only It blocks can be skipped
and even then they cannot list a reason - the tests merely pass. So
the tests for Get-DscResourceParameterInfoByCimClass only *actually*
run in an elevated context.