(MAINT) Ensure New-PuppetDscModule populates readme
Prior to this commit the call to Update-PuppetModuleReadme was
erroneously left out of New-PuppetDscModule, meaning that the
README that shipped with each auto-generated module was the
default template readme supplied by the PDK instead of the
useful and specific one generated for each DSC module.

This commit adds the call and updates tests to reflect it.