(IAC-740) Map embedded CIM instances if running elevated
This commit plumbs the Get-DscTypeInformation function to be aware
of whether or not it is running in an elevated context or not and
to retrieve the parameter information by cim class if it *is*
running elevated, or by AST/Resource metadata if not.

This also updates the Get-TypeParameterContent function to more
neatly display the structs for embedded instance types, ensuring
their indentation makes them (relatively) easy to read and parse.
This only matter if they can be mapped, which is only possible
when running elevated.

This commit adds a warning to the New-PuppetDscModule function,
alerting users to the fact that it will not be able to mapa any
embedded CIM instances unless re-run with Administrator permissions.

Finally, this commit add the Test-RunningElevated private function
which is just a wrapper around .NET methods to make it possible to
test functionality without having to actually run the tests elevated
and unelevated alike.