(IAC-846) Canonicalize Puppet module author names
Prior to this commit it was possible for a user to specify a
Puppet module author name as a string which was not a valid
name - that is, included non-alphanumeric characters and/or
uppercase characters.

This commit adds a private function for canonicalizing the
author name and ensures that the author name is canonicalized
when called for creating a new module.