(MAINT) Distinguish between key and required properties
Prior to this commit the builder did not distinguish between DSC Resource
properties which were mandatory to execute an invocation and thosee which
defined the resource uniquely on the system.

This information is *not* retrievable via the metadata returned from the
Get-DscResource command, only when parsing the CIM instances.

This commit adds an additional key to the ParameterInfo object and parses
the flags on CIM instances of DSC resources to determine whether the
property is a namevar (has the Key flag) or merely mandatory (has the
Required flag). It sets a value in the output hash for is_namevar, which
is used in the Get-TypeParameterContent function to determine whether or
not to add the namevar behavior.

The CIMless implementation takes the shortcut of assuming tha any
property which is mandatory is also a namevar; this is not ture and may
cause problems during runs, but we cannot distinguish any further with
the available information surfaced by the Get-DscResource command or the
AST.