(MAINT) Fix and format preamble script
Prior to this commit the preamble script did not use correct PowerShell formatting.
Furthermore, the canonicalization of nested CIM instances was inaccurate, leading
to churn on applies.

This commit corrects the formatting and ensures that nested CIM arrays are returned
properly canonicalized, with the appropriate data structures and cim_instance_type.