(IAC-741) Do not downcase values by default
Prior to this commit all values during both build/parse and canonicalization
after a get request were downcased.

This commit ensures that the downcasing is no longer performed, enabling the
tooling to take advantage of PascalCased default values, camelCased keys for
nested CIM instances, and prepare for canonicalization in the provider which
enables case-insensitive but case-preserving behavior.