(MAINT) Fix get/set invocation behavior
Prior to this commit the get and set invocation methods had naive behavior
for passing properties to Invoke-DscResource.

This commit ensures that the only properties passed for a get operation are
those mandatory for that operation, discarding all others.

For set operations, it rejects all of the Puppet-centric attributes which
do not need to be passed to PowerShell (and which cause errors if passed).

Finally, this commit clears instantiated variables after each assembling
of the script content to keep the debug information manageable.