(IAC-741) Implement Canonicalize feature
Prior to this commit Puppet had to treat dsc resource behavior
as case sensitive. This commit implements the Resource API
canonicalize feature and method to conform the values in the
manifest to those found on disk, if necessary, which ensures
that the manifests are written case insenstively but also
that case is preserved.

The canonicalize method itself is called first against the
resources defined in the manifest and then again after the
provider retrieves any resources from the machine.

The method has to distinguish between the canonicalization
of the manifest and of the retrieved values, so this commit
also implements a cache; the cache is empty when run against
the manifest resources and not empty when checking values
from the machine.

This commit also updates the get method, ensuring that any
mandatory values specified in the manifest but which are
not themselves namevars *are* passed to invoke_get_method,
preventing errors during execution.

Finally, this commit modifies the invoke_get_method slightly,
conforming the returned objects to be comparable and correct.