(IAC-895) Implement query caching
Prior to this commit the get method was invoked for a unique resource
regardless of whether or not that information had already been
retrieved in a prior query or not.

This commit adds a new cache for query results to reduce unneccesary
operations, thus speeding things along and reducing resource
utilization.

This commit also uses a new method, fetch_cached_hashes, to verify
if a given set of hashes can be looked up in a cache; in other words,
if the value of a passed hash is a subset of any of the cached hashes,
the method will return the full value of the cached hash.

This new method has been used for both the canonical and query caches.

Further, this commit updates the behavior of the get method to only
invoke if the name hash for the resource has not already been queried;
otherwise it shortcircuits and simply returns the cached value.

Finally, this commit makes minor updates to the handling of the ensure
and dsc_psdscrunascredential values to prevent flapping with the new
caching mechanism in place.