(IAC-1045) Extricate base provider
Prior to this commit the code for the base provider of the new
puppetized DSC modules was vendored into each module; this
meant that if the provider ever needs to be updated, every
prior release of the puppetized module would need to be built
again and pushed with an incremented number.

Moving the base provider to the pwshlib Puppet module means
that the provider can be updated separately from the PowerShell
builder code and does not require rebuilding puppetized modules,
only updating the reference pin for pwshlib.

This also makes testing the provider easier as it lives in a
ruby library now instead of in a PowerShell module.

This commit completes the extrication on the builder side by
removing the templated files and updating the provider definition
to point at the new location.