(MAINT) Parameterize unit test output file
Prior to this commit the unit test output file destination was
only set in the General Testing section, meaniing tests of the
functions-only would error out.

This commit moves that definition into the param block, meaning
that the default path is available regardless of the tests that
are being run; this also makes that valuable overrideable.