(MAINT) Enable puppetizing prerelease modules
Prior to this commit specifying a module with a prerelease version
would cause an error because the AllowPrerelease flag was not set.

This commit merely enables the puppetizing of prerelease modules but
does not rework the version API to distinguish between them. Because
of the way the modules are currently built, Puppet will not publish
prerelease modules to the forge.

While prerelease versions *can* be used, the maintainers suggest that
this is always a temporary workaround and not a good practice to be
encouraged in production.