(GH-66) Ensure type desc strings parse for reference
Prior to this commit the desc strings for types were wrapped in %q{},
which *should* allow multi-line string description blocks. However,
there is a bug in the puppet strings tooling which causes it instead
to write the %q{} as well as everything inside, which looks _strange_
in the markdown.

This commit modifies the type generation to instead use a single-
quoted string for the desc blocks, without newlines. If the help is
empty, it now specifies a string with a single whitespace in it.

Failure to include the single whitespace broke puppet strings parsing,
so this ensures that no docs are written if they are not specified,
keeping the reference documentation readable if not correct from a
markdown-linting perspective.

- Resolves #66