(GH-71) Fix type def of props with optional nested cim instances
Prior to this commit nested CIM instances would get an incorrect Puppet
type definition, where optional types would not be prefaced as such.

This commit updates the section of Get-DscResourceParameterInfoByCimClass
which handles the Puppetization of the data type for nested CIM instances,
adding a final check to validate whether or not the property is actually
mandatory. This ensures that the types are correctly defined and can be
passed as expected in a manifest.