(GH-74) Remove ensure keyword
This commit removes the ensure keyword from the generated types,
leaving only the dsc_ensure keyword *if* the resource requires
it - this moves towards treating resources as ensurable if and
only if they actually are. It also removes the UX confusion that
surrounds having both the ensure and dsc_ensure keywords but
requiring users to only ever specify dsc_ensure... unless they
want to ensure it is absent, requiring them to specify both.

- Resolves #74