(GH-81) Handle parameters in type generation
This commit adds handling for treating non-retrievable
DSC Resource properties as Puppet parameters, including
creating a new Test-DscResourcePropertyParameterStatus
private function which is utilized in the Get-ParameterInfo
functions. The value of this test is checked in the
Get-TypeParameterContent function to write the appropriate
behaviour in the generated type definition.

Depends on the implementation in puppetlabs/ruby-pwsh#62.

Resolves #81