(GH-94) Add debugging env to vagrant
This commit updates the vagrant provisioning to prepare for the
installation of rubies via uru/rubyinstaller and improves the
dev ux in the vagrant box by adding posh-git and a helper function
for creating a new Gemfile.local with the pry gems.

This should enable pry debugging during generated module testing.

NB: The automatic ruby installer is non-functional, users will need
to run Install-Ruby interactively when first logging into the new
Vagrant box.

- Resolves #94