(GH-85) Make puppetizing prerelease modules switchable
Prior to this commit the AllowPrerelease flag was always passed
to Save-Module in the Add-DscResourcModule function. This meant
that if no version was specifically passed during a call to
New-PuppetDscModule or Add-DscResourceModule they would always
default to the latest version on the Gallery *including* pre-
release versions which is unexpected UX.

This commit ensures that prerelease versions are only puppetized
if the AllowPrerelease switch is specified and adds that parameter
to both functions.