(GH-89) Ensure Get-PowerShellDscModule helper function can find xPhp
The xPhp module does not correctly indicate that it has DSC Resources to export,
which causes the logic in Get-PowerShellDscModule to be unable to locate it, even
when specifying by name.

This commit ensures that modules specified by name to be found and puppetized will
be attempted even if their metadata is incorrect.

This will NOT effect functionality when no name is specified.