(GH-113) Vendor modules not found in module repo
Prior to this commit some modules, like the PrintManagementDsc module,
include dependencies in the RequiredModules key of their manifest which
cannot be found on the public gallery. This caused errors during the build
process as the ModulePath is temporarily munged.

This commit overcomes that problem by vendoring the dependencies in, even if
they cannot be found in the specified repo, so long as they are available on
the machine puppetizing the PowerShell module containing DSC Resources.

- Resolves #113