(GH-92) Convert Tests to Pester 5
This commit moves the tests from Pester 4 to Pester 5; it also moves
the tests towards the idiomatic placement of tests, ie, side-by-side
with the functions they test.

Furthermore, this removes some dead test code and brings us in line
for more maintainable expansion of our tests in the near future.

Minor edits have been made to the PSM1 to avoid attempting to load the
new test files on module import.