(GH-93) Allow alternate fixture references
Prior to this commit, the generated Puppet module fixture yaml would always
include a reference to the latest released version of puppetlabs-pwshlib on
the Forge. The meant that any testing of an alternate version - an older
release or a current git branch - required manual intervention which itself
needed the user to know and understand Puppet spec fixtures.

This commit modifies the Update-PuppetModuleFixture private function to be
able to accept hashtables which define the fixture reference and do so in
a way that is idiomatic to the underlying implementation.

It includes a default behavior which matches the extant behavior, ensuring
that this change is backward compatible, and documentation for the new
behavior (including an example of use).

Further, it adds basic validation to the fixture hash, ensuring that what
is written is as valid as we can make it - we have no way to know without
running additional code if the git references or branch are valid.

This commit does NOT expose the new functionality in the public functions.