(GH-93) Ensure Acceptance tests run vs release & main
Prior to this commit the acceptance tests always and only ran against the latest
released version of pwshlib on the Forge. This commit splits the acceptance tests
into two cells: one for the latest release on the Forge, one for the latest changes
on the main branch of the GitHub repository.

This can be modified for testing a PR by altering the data in the acceptance-github
configuration (lines 102-106), by changing the repo or branch definition or adding
a ref key for a specific commit SHA.