(GH-133) Throw if PSRemoting is disabled
Prior to this commit, the New-PuppetDscModule command could fail unexpectedly
if PSRemoting is not enabled; when attempting to map CIM instances the command
would fail with an esoteric null error.

This commit adds an explicit check to the command which validates that PSRemoting
is enabled if running as administrator and errors clearly if it is not.

- Resolves #133