(GH-145) Add validation mode flag to puppetized DSC Resources
This commit adds a validation mode flag to the puppetized resources which
changes their behavior from Puppets perspective; by default resources will
use the property validation mode, which relies on Puppets own property
insync validation to tell whether or not a resource is in the desired
state. If the manifest writer instead specifies the validation mode to be
resource mode, it will call Invoke-DscResource with the test method and
trust that result, reporting in/out of state for **all** properties of
the resource. This results in less granular change reporting but reduces
the incidence of flapping, failing to recognize changes correctly, etc.

This commit adds documentation for using the validation mode flag and
tests to validate the changes in behavior.

This change is blocked on the resolving of puppetlabs/puppet-resource_api#225
and its subsequent release into Puppet, as well as a PDK version which
bundles those versions into itself.

It is also blocked on the merging of puppetlabs/ruby-pwsh#124 which implements
the custom insync logic in the base provider.

TODO:

- [ ] Write a changelog entry
- [ ] Ensure tests function as expected