(MAINT) Remove unused test files
Prior to this commit, tests written with the  Pester v4 syntax
were kept in the src/tests/ folder; since the move to Pester v5,
unit tests for functions have been moved to their idiomatic
location beside the functions they test. Similarly, the interface
script, Pester.ps1, is no longer used as it is incompatible with
the v5 syntax.

This commit removes these unused and unneeded files.