(MAINT) Move acceptance tests to folder
This commit moves the acceptance tests to their own folder in preparation
for adding multiple acceptance test files. In the current design, all
acceptance tests would need to live in a single file which makes it hard
to maintain and reason about.

This commit moves the existing acceptance tests to acceptance/Basic.Tests.ps1
and updates the CI config to look in the correct place.