(MAINT) Refactor unit/general jobs
This commit refactors the unit and general jobs in GHA which shared setup
and execution steps with minor differences to be a single job with a more
descriptive matrix.

The new spec job manages which tests to run using tags and sets the result
file name appropriately to the tests being run using the same matrix.

A future commit should refactor the acceptance tests to also use this new
job, further reducing the amount of reused code in the workflow.