(MAINT) Make acceptance actually work against git/forge
Prior to this commit, despite the use of fixtures to override the
source of puppetlabs/pwshlib, the tests always and only used the
latest released version.

This happened because all of the apply tests worked against the
*installed* modules - and when installing a generated module,
Puppet automatically installs the latest puppetlabs/pwshlib to
the Puppet module path.

This commit updates the basic acceptance tests to instead use
the module path in spec/fixtures/modules, which *does* utilize
the fixtures set up in the module build.

This ensures that acceptance tests in CI actually run against
both the latest release of puppetlabs/pwshlib as well as the
main branch of the ruby-pwsh repo.