(GH-172) Add implementation to generated types
Prior to this commit, the generated types did not include their
implementation information.

This commit ensures that granular implementation information is added
to the `dscmeta_resource_implementation` key, distinguishing between
MOF-based, class-based, binary, and composite implementations instead of
only Powershell/binary/composite.

To do so, this commit:

1. Adds the `Get-DscResourceImplementation` private function, which has
   logic to determine if a DSC Resource with an `ImplementedAs` value of
   `PowerShell` is MOF-based or class-based
2. Ensures that the `Get-DscResourceTypeInformation` private function
   uses `Get-DscResourceImplementation` to capture the implementation of
   the DSC Resource being introspected
3. Ensures that the `dscmeta_resource_implementation` key is added to
   the generated Puppet Resource API type definition for DSC Resources,
   using the value as determined in `Get-DscResourceImplementation`

This change is backwards compatible as it _adds_ a new key to the type
definition instead of changing any existing keys.

As referenced in puppetlabs/ruby-pwsh#173, this improvement will enable
the base provider to invoke class-based DSC Resources by munging the
system PSModulePath environment variable _only_ for those resources,
instead of for all resource invocations.

- Resolves puppetlabs/ruby-pwsh#172